# specs/001-build-a-comprehensive/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Amharic Document Preparation System API
  description: |
    Comprehensive API for processing, analyzing, and managing Amharic documents with multi-agent orchestration.
    
    ## Features
    - Multi-format document upload (PDF, images, Word, CSV, web content)
    - Real-time processing with CrewAI agents
    - Amharic-specific text analysis and OCR
    - Advanced search and export capabilities
    - Quality assurance and validation metrics
    
  version: 1.0.0
  contact:
    name: Amharic Document System Team
    email: support@amharic-docs.ai
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.amharic-docs.ai/v1
    description: Production server
...
paths:
...
  /documents:
    get:
      tags: [Documents]
      summary: List user documents
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DocumentStatus'
        - name: document_type
          in: query
          schema:
            $ref: '#/components/schemas/DocumentType'
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, filename, file_size, processed_at]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'

  # Processing Endpoints
...
  /search/documents:
    post:
      tags: [Search]
      summary: Search documents
      description: |
        Full-text search across processed documents with Amharic language support.
        Supports both keyword and semantic search with faceted filtering.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/suggestions:
    get:
      tags: [Search]
      summary: Get search suggestions
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        '200':
          description: Search suggestions retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # Export Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        filename:
          type: string
        document_type:
          $ref: '#/components/schemas/DocumentType'
        file_size:
          type: integer
        content_hash:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        metadata:
          $ref: '#/components/schemas/DocumentMetadata'
        created_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
    DocumentType:
      type: string
      enum: [pdf, image, word, csv, web_content]
    DocumentStatus:
      type: string
      enum: [uploaded, queued, processing, completed, failed, archived]
    DocumentMetadata:
      type: object
      properties:
        original_name:
          type: string
        mime_type:
          type: string
        page_count:
          type: integer
        language_detected:
          type: string
        creation_date:
          type: string
          format: date-time
        author:
          type: string
    DocumentList:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # Processing Schemas
...
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query string
        filters:
          $ref: '#/components/schemas/SearchFilters'
        sort:
          $ref: '#/components/schemas/SearchSort'
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
    SearchFilters:
      type: object
      properties:
        document_types:
          type: array
          items:
            $ref: '#/components/schemas/DocumentType'
        languages:
          type: array
          items:
            type: string
        date_range:
          type: object
          properties:
            from:
              type: string
              format: date
            to:
              type: string
              format: date
        quality_score:
          type: object
          properties:
            min:
              type: number
              format: float
            max:
              type: number
              format: float
    SearchSort:
      type: object
      properties:
        field:
          type: string
          enum: [relevance, date, title, quality_score]
          default: relevance
        order:
          type: string
          enum: [asc, desc]
          default: desc
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        facets:
          $ref: '#/components/schemas/SearchFacets'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        total_matches:
          type: integer
        query_time:
          type: number
          format: float
    SearchResult:
      type: object
      properties:
        document:
          $ref: '#/components/schemas/Document'
        relevance_score:
          type: number
          format: float
        highlights:
          type: array
          items:
            type: string
        snippet:
          type: string
    SearchFacets:
      type: object
      properties:
        document_types:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer
        languages:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              count:
                type: integer

    # Export Schemas
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...